improve unit test
